requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
PyMongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
python-dateutil==2.8.2
Werkzeug==2.3.7